Rename Parcel.FrontEnd.NodifyWPF to Parcel.Neo
