--- Parcel.FrontEnd.NodifyWPF/Converters/ProcessorNodeToNodeFullNameConverter.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Globalization;
-using System.Windows;
-using System.Windows.Data;
-using Nodify;
-using Parcel.Shared.Framework.ViewModels;
-using Parcel.Shared.Framework.ViewModels.BaseNodes;
-
-namespace Parcel.FrontEnd.NodifyWPF.Converters
-{
-    public class ProcessorNodeToNodeFullNameConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is ProcessorNode node)
-            {
-                return node.FullName;
-            }
-
-            return value;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
