+++ Parcel.Neo/App.xaml
@@ -1,8 +1,8 @@
-﻿<Application x:Class="Parcel.FrontEnd.NodifyWPF.App"
+﻿<Application x:Class="Parcel.Neo.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Parcel.FrontEnd.NodifyWPF"
-             xmlns:converters="clr-namespace:Parcel.FrontEnd.NodifyWPF.Converters"
+             xmlns:local="clr-namespace:Parcel.Neo"
+             xmlns:converters="clr-namespace:Parcel.Neo.Converters"
              StartupUri="MainWindow.xaml"
              Startup="ParcelApplicationInitialize">
     <Application.Resources>
