+++ Parcel.Neo/BaseWindow.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Windows;
 
-namespace Parcel.FrontEnd.NodifyWPF
+namespace Parcel.Neo
 {
     public class BaseWindow: Window, INotifyPropertyChanged
     {
