+++ Parcel.Neo/CommentWindow.xaml.cs
@@ -6,7 +6,7 @@ using ICSharpCode.AvalonEdit.Highlighting;
 using Parcel.Shared.Framework;
 using Parcel.Shared.Framework.ViewModels.BaseNodes;
 
-namespace Parcel.FrontEnd.NodifyWPF
+namespace Parcel.Neo
 {
     public partial class CommentWindow : BaseWindow
     {
