+++ Parcel.Neo/Converters/HiddenVisibilityConverter.cs
@@ -3,18 +3,17 @@ using System.Globalization;
 using System.Windows;
 using System.Windows.Data;
 using Nodify;
-using Parcel.Shared.Framework;
 using Parcel.Shared.Framework.ViewModels;
 
-namespace Parcel.FrontEnd.NodifyWPF.Converters
+namespace Parcel.Neo.Converters
 {
-    public class NodeMessageTypeToVisibilityConverter : IValueConverter
+    public class HiddenVisibilityConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is NodeMessageType messageType)
+            if (value is bool isHidden)
             {
-                return messageType != NodeMessageType.Empty ? Visibility.Visible : Visibility.Collapsed;
+                return isHidden ? Visibility.Collapsed : Visibility.Visible;
             }
 
             return value;
@@ -24,7 +23,7 @@ namespace Parcel.FrontEnd.NodifyWPF.Converters
         {
             if (value is Visibility visibility)
             {
-                return visibility == Visibility.Collapsed ? NodeMessageType.Empty : NodeMessageType.Normal;
+                return visibility == Visibility.Collapsed ? true : false;
             }
 
             return value;
