+++ Parcel.Neo/MainWindow.xaml
@@ -1,14 +1,14 @@
-﻿<local:BaseWindow x:Class="Parcel.FrontEnd.NodifyWPF.MainWindow"
+﻿<local:BaseWindow x:Class="Parcel.Neo.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Parcel.FrontEnd.NodifyWPF"
+        xmlns:local="clr-namespace:Parcel.Neo"
         xmlns:nodify="clr-namespace:Nodify;assembly=Nodify"
         xmlns:viewModels="clr-namespace:Parcel.Shared.Framework.ViewModels;assembly=Parcel.Shared"
         xmlns:baseNodes="clr-namespace:Parcel.Shared.Framework.ViewModels.BaseNodes;assembly=Parcel.Shared"
         xmlns:primitives="clr-namespace:Parcel.Shared.Framework.ViewModels.Primitives;assembly=Parcel.Shared"
-        xmlns:converters="clr-namespace:Parcel.FrontEnd.NodifyWPF.Converters"
+        xmlns:converters="clr-namespace:Parcel.Neo.Converters"
         xmlns:nodes="clr-namespace:Parcel.Toolbox.DataProcessing.Nodes;assembly=Parcel.Toolbox.DataProcessing"
         xmlns:System="clr-namespace:System;assembly=System.Runtime"
         xmlns:dataTypes="clr-namespace:Parcel.Shared.DataTypes;assembly=Parcel.Shared"
