+++ Parcel.Neo/PreviewWindow.xaml
@@ -1,9 +1,9 @@
-﻿<local:BaseWindow x:Class="Parcel.FrontEnd.NodifyWPF.PreviewWindow"
+﻿<local:BaseWindow x:Class="Parcel.Neo.PreviewWindow"
             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-            xmlns:local="clr-namespace:Parcel.FrontEnd.NodifyWPF"
+            xmlns:local="clr-namespace:Parcel.Neo"
             mc:Ignorable="d" MouseDown="PreviewWindow_OnMouseDown"
             Title="{Binding Node, Converter={StaticResource ProcessorNodeToNodeFullNameConverter}}" 
             Height="200" MinWidth="420" SizeToContent="Width"
