+++ Parcel.Neo/PropertyWindow.xaml
@@ -1,10 +1,10 @@
-﻿<local:BaseWindow x:Class="Parcel.FrontEnd.NodifyWPF.PropertyWindow"
+﻿<local:BaseWindow x:Class="Parcel.Neo.PropertyWindow"
                   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                  xmlns:local="clr-namespace:Parcel.FrontEnd.NodifyWPF"
-                  xmlns:converters="clr-namespace:Parcel.FrontEnd.NodifyWPF.Converters"
+                  xmlns:local="clr-namespace:Parcel.Neo"
+                  xmlns:converters="clr-namespace:Parcel.Neo.Converters"
                   xmlns:framework="clr-namespace:Parcel.Shared.Framework;assembly=Parcel.Shared"
                   mc:Ignorable="d"
                   Title="{Binding Processor, Converter={StaticResource ProcessorNodeToNodeFullNameConverter}}"
