+++ Parcel.Neo/README.md
@@ -1,3 +1,3 @@
-﻿# Parcel.FrontEnd.NodifyWPF
+﻿# Parcel.Neo
 
 A Parcel frontend written in WPF using Nodify as node graph framework.
